requirements.txt
pyyaml
jupyter
k-util
k-vision
opencv-python
pandas